Conflicts:
	build/vs2008/create_setup.nsi : removed
	build/vs2008/opencaesar3.vcproj : removed
	oc3_caesar.cpp : resolved
	oc3_scenario.cpp : resolved
	oc3_scenario.hpp : resolved
	oc3_screen_game.cpp : resolved
	oc3_screen_game.hpp : resolved
	oc3_terraininfo.cpp : resolved

Former-commit-id: 6b1f0917f22c766d232ed7fbfe140e73abe3e739

+++ oc3_city.cpp
@@ -469,7 +469,7 @@ void City::save( VariantMap& stream) const
   stream[ "boatExitI" ] = _boatExitI;
   stream[ "boatExitJ" ] = _boatExitJ;
   stream[ "climate" ] = _climate;
-  stream[ "time" ] = _d->time;
+  stream[ "time" ] = (unsigned long long)_d->time;
   stream[ "funds" ] = _d->funds;
   stream[ "populaton" ] = _d->population;
 //
